requirements.txt
Django==1.8
DjangoUeditor==1.8.143
django-filter==2.0.0
django-filters==0.2.1
django-pure-pagination==0.3.0
mysqlclient==1.3.13
PyMySQL==0.9.3
Pillow==5.2.0
pytz==2019.3
